[frontend] Making the system platform-agnostic
+ Made the proxy work in a separate way to proceed to implement back Netlify
