+++ frontend/netlify.toml
@@ -1,0 +1,14 @@
+[[redirects]]
+from = "/.ory/*"
+to = "https://romantic-satoshi-kojdtfzsl2.projects.oryapis.com/:splat"
+status = 200
+force = true
+[[redirects]]
+from = "/api/*"
+to = "https://calendarium-api.hopoke.workers.dev/api/:splat"
+status = 200
+force = true
+[[redirects]]
+from = "/*"
+to = "/index.html"
+status = 200
