+++ frontend/src/main.ts
@@ -33,7 +33,7 @@ router.beforeEach(async () => {
                 logoutUrl.value = data.logout_url.replace("https://romantic-satoshi-kojdtfzsl2.projects.oryapis.com/", "https://"+window.location.origin+"/.ory/")
             })
         } catch (e) {
-                window.location.href = "/.ory/ui/login" + (vercelEnv == 'development' ? '' : '?return_to='+window.location.origin)
+            window.location.href = "/.ory/ui/login" + (vercelEnv == 'development' ? '' : '?return_to='+window.location.origin)
         }
     }
 })
